+++ doc/whats_new.rst
@@ -19,7 +19,12 @@ Current
 Changelog
 ~~~~~~~~~
 
-- Add :class:`mne.SourceMorph` with :meth:`mne.SourceMorph.save` and :meth:`mne.SourceMorph.as_volume`, :func:`mne.read_source_morph`, by `Tommy Clausner`_
+- Add :class:`mne.SourceMorph` with :meth:`mne.SourceMorph.save` and :meth:`mne.SourceMorph.as_volume`, :func:`mne.read_source_morph`, 
+by `Tommy Clausner`_
+
+- Deprecated :meth:`mne.VolSourceEstimate.morph_precomputed` 
+and :func:`mne.morph_data_precomputed` in favor of 
+:class:`mne.SourceMorph` by `Tommy Clausner`_
 
 - Add possibility to save :class:`mne.VolSourceEstimate` and :class:`mne.MixedSourceEstimate` to HDF5 format (file extension .h5) with :meth:`mne.VolSourceEstimate.save` and :meth:`mne.MixedSourceEstimate.save` by `Alex Gramfort`_